Move part of StateExplorerUtils to vespalib.
