+++ vespalib/src/vespa/vespalib/datastore/data_store_explorer.cpp
@@ -4,23 +4,11 @@
 #include "datastorebase.h"
 #include <vespa/vespalib/data/slime/cursor.h>
 #include <vespa/vespalib/util/memoryusage.h>
+#include <vespa/vespalib/util/state_explorer_utils.h>
 
 using vespalib::slime::Cursor;
 
 namespace vespalib::datastore {
-
-namespace {
-
-void
-memory_usage_to_slime(const MemoryUsage& usage, Cursor& object)
-{
-    object.setLong("allocated", usage.allocatedBytes());
-    object.setLong("used", usage.usedBytes());
-    object.setLong("dead", usage.deadBytes());
-    object.setLong("onHold", usage.allocatedBytesOnHold());
-}
-
-}
 
 DataStoreExplorer::DataStoreExplorer(const DataStoreBase &store)
     : StateExplorer(),
@@ -35,7 +23,7 @@ DataStoreExplorer::get_state(const vespalib::slime::Inserter& inserter, bool ful
 {
     (void) full;
     auto& object = inserter.insertObject();
-    memory_usage_to_slime(_store.getMemoryUsage(), object.setObject("memory_usage"));
+    StateExplorerUtils::memory_usage_to_slime(_store.getMemoryUsage(), object.setObject("memory_usage"));
     auto bufferid_limit = _store.get_bufferid_limit_acquire();
     auto max_num_buffers = _store.getMaxNumBuffers();
     object.setLong("bufferid_limit", bufferid_limit);
